\name{matrix_unknown}
\alias{matrix_unknown}
\docType{data}
\usage{data("matrix_unknown")}
\title{
Matrix with 4 unknown Raman spectra of plastic polymers
}

\description{
Database with frequency data as a first column ("freq"), and intensity values of 4 different unknown plastic polymers (purely by way of example).
}
\examples{
data("matrix_unknown")
str(matrix_unknown)
summary(matrix_unknown)
}
\keyword{dataset}
