\name{MPdatabase}
\alias{MPdatabase}
\docType{data}
\usage{data("MPdatabase")}
\title{
Database with Raman spectra of plastic polymers and pigments
}

\description{
Database with frequency data as a first column ("freq"), and intensity values of different plastic polymers and plastic additives.
}
\examples{
data("MPdatabase")
str(MPdatabase)
summary(MPdatabase)
}
\keyword{dataset}
