\name{single_unknown}
\alias{single_unknown}
\docType{data}
\usage{data("single_unknown")}
\title{
Matrix with 1 unknown Raman spectra of plastic polymer
}

\description{
Database with frequency data as a first column ("freq"), and intensity values of 1 unknown plastic polymers (purely by way of example).
}
\examples{
data("single_unknown")
str(single_unknown)
summary(single_unknown)
}
\keyword{dataset}
